requirements.txt
motor==3.3.2
python-dotenv==1.0.1
pymongo==4.6.1
psutil==5.9.8
aiohttp==3.9.1
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
certifi==2024.2.2
dnspython==2.4.2
orjson==3.9.9
pyopenssl==24.0.0
prometheus-client==0.19.0
pydantic==2.5.0